requirements.txt
github3.py==0.9.3
ipdb==0.8
ipython==2.3.1
py==1.4.30
pytest==2.7.2
requests==2.7.0
uritemplate.py==0.3.0
